fix(better-auth): extract DI tokens to break module/service circular import (SWC TDZ crash)
The better-auth DI tokens were split across two files that imported each
other: BETTER_AUTH_INSTANCE was declared in core-better-auth.module.ts,
while BETTER_AUTH_CONFIG and BETTER_AUTH_COOKIE_DOMAIN were declared in
core-better-auth.service.ts. The service imported BETTER_AUTH_INSTANCE
from the module, and the module imported the other two tokens (plus the
service class) back from the service — a genuine import cycle.

Under tsc/CommonJS this happened to work by evaluation-order luck. Under
SWC (`nest start -b swc`) it crashes at startup:

  ReferenceError: Cannot access 'BETTER_AUTH_INSTANCE' before initialization

The tokens are dereferenced at module-evaluation time (`@Inject(...)` in
the CoreBetterAuthService constructor is evaluated when the class is
defined), so the cycle lands in the temporal dead zone.

Fix: move all three tokens into a dedicated leaf module,
core-better-auth.constants.ts, which imports nothing. Module and service
both import the tokens from there, making the graph acyclic and the
initialization order deterministic for every compiler.

Backward compatible: both core-better-auth.module.ts and
core-better-auth.service.ts re-export the tokens, so existing deep
imports (`from '.../core-better-auth.module'` /
`'.../core-better-auth.service'`) and the package barrel keep working
unchanged. No public API change.

Verified by compiling src/ with SWC (legacy decorators + decoratorMetadata,
matching the Nest CLI's SWC builder) and requiring the module: the
ReferenceError reproduces before this change and is gone after it. madge
reports the better-auth cycles dropping from 9 to 1; the remaining
guard -> module cycle is benign because it only dereferences
CoreBetterAuthModule lazily inside a method body, never at eval time.

Build, lint, format, 691 unit tests and 1380 e2e tests all pass.
